refactor(svelte): extract interaction diagnostic catalog module
Move INTERACTION_DIAGNOSTIC_CATALOG and its types out of interaction.ts so
pure agent diagnostic data is not mixed with event types and config
normalization. Public imports stay on interaction.js via re-exports.
Characterization locks each catalog entry's code field to its Record key.
